create a generic deleter
A custom deleter for std::{unique,thread}_ptr is added. To implement the
deletion strategy for a type, the annihilate() function must be
specialized. As an example, suppose that we have the following API
exposed by a C library:

    struct Foo { ... };
    Foo *MakeFoo();
    void DestroyFoo(Foo *foo);

but we don't know how Foo objects are allocated by MakeFoo(), so we
have to use always DestroyFoo() instead of free()/delete. In order to
make a smart pointer to Foo, first specialize the annihilate() function
like this:

    template<>
    void annihilate<Foo>(Foo *foo)
    { DestroyFoo(foo); }

and then we can use the Annihilator class as the deleter:

    std::unique_ptr<Foo, Annihilator<Foo>> ptr { MakeFoo() };

This guarantees that the object will be destroryed using the proper
function.